backend/requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Configuration
pydantic>=2.9.0
pydantic-settings>=2.6.0

# Database (async PostgreSQL + pgvector)
asyncpg>=0.30.0
sqlalchemy[asyncio]>=2.0.36
pgvector>=0.3.6
alembic>=1.14.0

# Scientific / vectors
numpy>=1.26.0

# HTTP client
httpx>=0.27.0

# Google Cloud
google-cloud-storage>=2.18.0
google-cloud-aiplatform>=1.72.0
google-cloud-tasks>=2.16.0
google-genai>=1.0.0

# PDF extraction
PyMuPDF>=1.24.0

# PDF generation (HTML → PDF)
Jinja2>=3.1.0
weasyprint>=62.0
pypdf>=4.0.0

# LangGraph / LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-vertexai>=2.0.0
